requirements.txt
attrdict==2.0.1
imageio==2.6.1
matplotlib~=3.5.0
numpy~=1.21.4
Pillow
scikit-image~=0.16.2
scipy~=1.7.3
tqdm~=4.42.1
yacs~=0.1.8
tensorboard==2.1.1
rtpt~=0.0.4
argparse~=1.4.0
gym~=0.17.2
scikit-video~=1.1.11
mushroom_rl==1.6.1
scikit-video
filterpy~=1.4.4
torch~=1.10.0
opencv-python~=4.3.0.36
torchvision~=0.11.1
pandas~=1.0.1
scikit-learn~=1.0.1
termcolor~=1.1.0
ipython~=7.12.0
seaborn~=0.11.1
setproctitle~=1.2.2
filterpy~=1.4.5
sklearn~=0.0
